Fix telegram markdown

+++ Tadmor.Core/ChatClients/Telegram/Interfaces/ITelegramGuildUserFactory.cs
@@ -5,7 +5,7 @@ namespace Tadmor.Core.ChatClients.Telegram.Interfaces
 {
     public interface ITelegramGuildUserFactory
     {
-        Task<ITelegramGuildUser?> CreateOrNullAsync(ITelegramGuild telegramGuild, int userId);
+        Task<ITelegramGuildUser?> CreateOrNullAsync(ITelegramGuild telegramGuild, long userId);
         ITelegramGuildUser Create(ITelegramGuild telegramGuild, ChatMember chatMember);
         ITelegramGuildUser Create(ITelegramGuild telegramGuild, User user, bool isAdmin);
     }
